requirements.txt
eyed3
librosa
ffmpeg-python
python-vlc
gdown